requirements.txt
pgcli
psycopg2-binary
kafka-python
requests
timeloop
